src/requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py
discord_colorize==0.0.8
discord.ext.modal_paginator==1.1.1
names==0.3.0
tabulate==0.9.0
mariadb==1.1.8
